Write directly in buffer for alternate ft_printfs
This change concerns ft_sprintf, ft_snprintf and ft_asprintf functions.
From now on, they will be writing directly in the buffer given by the
user. This should make these functions faster to be executed.

At this stage, the project has only been tested and approved on
compilation. The functions modified by this commit should be further
tested before this branch is merged with master.