Add ft_wtlen and ft_wtdup
These functions act on "word tables" which are NULL terminated array of
strings (char **), like argv or env for example. ft_wtlen acts as
ft_strlen, by counting the number of strings before NULL, and ft_wtdup
copies an entire array of strings. It copies each string individually
and places them in an array of the same size as the original.