Collect every memory allocation and "objects"
This update of the libft finally adds ft_secmalloc, which secures
malloc calls by testing them and exiting cleanly on failure. This
is achieved with ft_heap_collector which stores every allocated
pointers in the program (except for the heap_stack which is "manually"
freed by the collector when giving it the instruction "FT_COLLEC_FREE").

It also adds the ft_collector function which is a generic version of
ft_heap_collector, to handle every other type of data/object which
needs to be properly closed/cleaned before exiting the program.

And finally, to complete the all thing, a set of exit functions has
been added. This is ft_exitmsg, ft_atexit and ft_exit. ft_exitmsg sets
the begining of the error message when leaving the program (this
typically should be its name). ft_atexit stores functions to call at
the end of the program (they will be executed in order LIFO). ft_exit
will print an error message if need be, call ft_atexit and simply
exit the program with the value given as a parameter.

Everything this update adds can be ignored at compile time by adding
-DNO_COLLEC to the C flags in the Makefile. When doing this, collectors
are not compiled and ft_secmalloc is just a regular malloc. The linker
shows "warnings" because some of the files given don't have symbols.
This is normal and shouldn't prevent the library from working, but this
should be silenced somehow (TODO).

This is not yet tested. This should be done with ft_ls and a good leak
detector.