Collect open directories with ft_dir_collector
This will allow for internal handling of open directories in the libft.
Next, listing functions will be added, and ft_dir_collector will be used
if NO_COLLEC is not set. Also, ft_close_dir_stack must be added to the
ft_atexit stack if these functions are to be used in a program with the
libft collector on.