Replace free call by ft_memdel to Normify
This is a bit shitty, but this works. In an other function I would
not have done that because it could reduce performance if the function
is called a lot (the effect would be almost unnoticeable), but since
this function is only called once if the ft_getopt fails, the effect
is negligeable.