Fix BINT_SET_SIGN and bint_sub_abs
Just put parenthesis around the sign so that an eventual ternary
expression would not change the result (like in this case where it
was putting the info block a 0 because of the nested ternary).

Also fix the length check in the bint_sub_abs. The function now
uses bintcmp_abs because that's the info that's really wanted, not
only the length of the number.