Cast the bint_command function pointer
Now it actually works, but the code is pretty messy tbh. The next
step is to refactor the entire thing. To add the useful new
functions to the bint library and to create a directory for the
bintbc specific code.

Once it is done, we can test the lib more in depth and start to add
the new functions needed for the ft_ssl project.