Fix 0 bug in divide_bint
When the substractions reached 0, BINT_LEN(res) was not set to 0
which is a problem because that's how 0 is represented for bint
numbers (it still mostly worked, but this was not it was supposed
to work). It was a problem for dragon4 because the loop condition
where BINT_LEN(scaled_val) == 0 was never valid, so the code after
(in efg conversions for ft_printf) did not work with the new
buffer output by dragon4. It was mission all the trailing 0s up to
cutoff_exp. This needs to be fixed too.