Rewrite ssub_u32_abs to use internal_bint_sub
This is more efficient and secure like this. Also, internal_bint_sub
seems to work well with res == large || res == small. So it should
be easy to make a bint_ssub.

NOTE: There is a problem with bint_print in base 10. I don't know
why but with 10000 - 50 == 9950 bint_print shows 99500. The hex
is correct so the value is right. This is probably due to the
buffer filling with 0s.