Remove bintinit and optimize bintclean and bintclr
Essentially, now the bints are officially not required to have
every digit set to 0 after the last digit (at BINT_LEN(b)). This
avoids reseting them all each time we init or clear a bint.

This is an optimization. Now modexp is 4 times as fast as it used
to be. Although, this is still not perfect. bint_mult_internal is
now the function that takes the longest to execute. There must be
a faster multiplying algorithm than the one I use (which is really
vanilla actually, so it should not be hard to improve on it).

Also tests have been added for modexp with its corresponding
modexpTests.c and modexpTests.h files. For now the tests with
8192 as a modulo and exponent dont yep work.

Maybe this is actually not a problem since what is really driving
the computation time up in modexp is the b parameter (the exponent)
since it loops on each of its bits. In practice, for the RSA
algorithm the exponent should be smaller than the key and so take
way less time to compute. Hopefully at least...