Implement modmul and add it to bc
The function is really simple and seems to work, although tests
still have to be written for it. This is the last modular function
needed for the RSA algorithm. After the tests are done for this one
an optimization of the multiplication algorithm could be looked for
as it pretty heavily brings the exec time up. After that, there
will only be the prime functions to implement.

Once all of this is done, and after a little bit more tests are
performed and written for the bint library, it will be possible to
actually use it for RSA. So ft_ssl will have to update its libft
version to the newest one and general bint support will have to be
implemented. It should not be too hard as the bint functions are
basically the same as uint64_t functions. The hardest thing will
probably be to respect the 25 lines limit when adding SUCCESS tests
for bint functions.