Set test cases up to 512 bits for mod ops
This is the farthest I can go without waiting too much for
the modexp operation. I think I will use the long_multiplication
algorithm and stop at 512 bits for the key size (or maybe more
with an option...). Anyway it conceptually works, this is only an
optimization problem now and need to go on.