Add bint primality functions
It works but it is sometimes very slow including for low size
values. The bint_find_prime, bint_is_prime and bint_miller_test
functions can probably be optimized a bit, but I think the main
problem is the repeated calls to modexp. Sometimes on "lucky"
values the number of modexp ops to do is quite low because we
are not far from a prime number. So my guess is that the bint
multiplication function is still at the root of the problem.

If it cannot be fixed otherwise, even by reducing the potential
size of the key to something disappointing like 256 or 128 bits,
the ft_ssl project will have to be able to generate 64 bit keys
only. Even in this case we can still use big integers for rsa
and rsautl, that way ft_ssl will be able to at least use bigger
keys even if it cannot generate them.