Replace useless MUTEXIFY calls by TLS variables
This way less costly just to ensure the static variable is local
to the thread. This is possible when the static variable is not
in a 'container' function or that the state does not include
resources already shared between threads (like file descriptors).
When the static is purely used in the function, like in
ft_split_whitespaces for example, it does not need to be protected
by a mutex as multiple threads wont need the information.