Remove ft_bufprintf
This was not very useful and it did not even work as intended. The only
purpose it could have had was to have a malloc-free printf, but that
does not even appear to have been the case. Also the behavior was not
coherent because it was flushing the buffer when the buffer was full and
it was also returning the buffer at the end, which means both could
happen and could mean the printing could be 'de-synchronized'.