Add ft_zprintf and ft_dzprintf functions
These functions are versions of printf that work entirely without
allocating space for the buffers it needs. It flushes its main buffer
regularly if it is too small for the content it should hold.

It looks like it's working well. The only thing is that it is limited by
the size of the static buffer. If a single conversion is big enough to
overflow the local buffer then it will STOP instead of flushing (and
then it should continue with other conversions in the format string).
Basically this means that gigantic conversions can be truncated for
extreme cases. This should not happen to often though, as zprintf and
dzprintf are designed for simple error/exit printing.

By the way z stands for zero, as in "zero allocation".

TODO: These function wont allocated memory in their internal logic, but
complex conversions could be different, especially for the float
conversions. The other ones are entirely implemented in ft_printf and do
not call extern functions that could allocated memory. Float conversions
use the dragon4 algorithm and the bint library at large. I'm pretty sure
it does not contain any dynamic memory allocation, but I'm not 100%
certain. This should be checked.