Make the ft_zprintf call fail if it has to stop
This actually makes more sense this way. Now when the static buffer is
not large enough to hold the current conversion at least the function
will notify the user by making the call fail.

Anyway, this is not really a big deal as this function is supposed to
run with limitations in specific instances. In practice this should not
happen too often and at worst it will lead to ugly printing and missing
newlines.