Create NOALLOC flag and add szprintf and snzprintf
They are versions of sprintf and snprintf with zero dynamic memory
allocation. The NOALLOC flag had to be added because there was no
difference between choosing flush and not reallocating. Meaning that we
could not have a 'bufferized' printf with zero allocation with the set
of flags we had before.

Now the pdata_add function, in case of a too small buffer, tries first
to STOP (only if the buffer is not local), if it is set to CONTINUE it
will try to FLUSH the buffer (once again, only if the buffer is not
local), else it will look at the NOALLOC flag. If it is set, whatever
the locality of the buffer, the ft_printf call will return an error.
Else, the pdata_realloc function will be called and the parsing will
continue as usual.

It is to be noted that the NOALLOC flag will be explicitly unset if
ALLOC is also present. ALLOC has precedence over NOALLOC because it is
used earlier in the code as the first bit is looked at to set the first
buffer at the start of the function.

It can be used with STOP or FLUSH because it also sets the behavior of
the local buffer and then it will be looked at.