feat: add packet filter functions
These functions will be useful for ping. Although a filter will need
to be made especially for it the general structure will make this easier.

TODO: document the kind of socket to be used with these functions or, even
better, create dedicated listener socket functions (maybe even use a typedef
for a 't_listener_socket' type that can be filtered with the ft_packet_filter
functions.