fix: compute icmpv4 checksum before sending ping
This works and we can see the reply is sent, although its not catched by the
ft_scan_poll function. Also, the checksum computation is little bit more
complicated for icmpv6 as it requires a part of the ipv6 header. If it's not
computed by the kernel it will be hard to fix because at this point in the
scan code, the ip header is not build because we don't have the src IP from
the local interface. I figured it was useless and simpler to just send from
the ICMP/TCP/UDP header, and I was right because it works well with IPv4. But
IPv6... I dont know, this will require a little bit of thinking if the kernel
does not do its job.