docs: remove outdated TODO comment
'hton' functions should not be used when setting the filters except for the
IP address. This is because apparently the kernel handles endianness on its
own when filtering packets except when it comes to the IP address (for the
ipv4 version anyway). This is a bit weird and completely empirical for now
but it works.