feat: list interfaces and setup network config
The network config is a purely internal libft structure (declared in the only
".c" file it is used in). It mainly consists of an array of 4 interfaces. Two
defaults and two loopbacks both on IPv4 and IPv6. This basically moves the
interface logic from ft_nmap to the libft. We need this because the tcp syn
scan needs a source IP to work (the goddamn kernel does not compute the TCP
checksum for outgoing packets through a SOCK_RAW/IPPROTO_TCP socket).