fix: build and send IP header for tcp syn scan
This really really sucks. But it works. To make a long story short, the
kernel does not compute the TCP checksum even when it builds the IP header
itself, which means it is always false. This means we can't rely on it to
choose the interface and only send the payload header (the TCP header in
this case and the ICMP header for echo scan). We have to do it ourself.